refactor: 재활용을 위한 ImageComponent 생성
